refactor: relocate Base58CryptoHash from near-crypto-hash to near-sdk-core
`Base58CryptoHash` (and its `ParseCryptoHashError`) is a JSON helper type
that only the SDK's `json_types` module consumes. Move its definition into
`near-sdk-core::json_types` — which already depends on bs58/serde_with/borsh
— so `near-crypto-hash` can shrink to just the bare `CryptoHash` alias.

The public path `near_sdk::json_types::Base58CryptoHash` is unchanged; only
`near-crypto-hash`'s own surface loses the type (breaking for that crate).
